% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{auc}
\alias{auc}
\title{Area Under the Curve}
\usage{
auc(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels. Can be \code{numeric, character, or factor}}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the area under the curve for a binary classifcation model
}
\examples{
data(testDF)
glmModel <- glm(y ~ ., data = testDF, family="binomial")
Preds <- predict(glmModel, type = 'response')

auc(testDF$y, Preds)

}

