% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{brier}
\alias{brier}
\title{Brier Score}
\usage{
brier(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the Brier score
}

