% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{error}
\alias{error}
\title{Average absolute error}
\usage{
error(actual, prediction, type = "absolute")
}
\arguments{
\item{actual}{A vector of the labels}

\item{type}{use \code{absolute} to calculate absolute error and \code{class} for classification error}

\item{predicted}{A vector of predicted values}
}

