% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{f1Score}
\alias{f1Score}
\title{F1 Score}
\usage{
f1Score(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{cutoff}{A cutoff for the predicted values}
}
\description{
Calculates the f1 score
}
