% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{f1Score1}
\alias{f1Score1}
\title{Mean F1 Score}
\usage{
f1Score1(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the mean f1 score
}

