% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{fScore}
\alias{fScore}
\title{F Score}
\usage{
fScore(actual, predicted, cutoff = 0.5, beta = 0.5)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{cutoff}{A cutoff for the predicted values}

\item{beta}{the desired beta value (lower increases weight of precision over recall)}
}
\description{
Calculates the F score and allows different specifications of the beta value (F0.5)
}
