% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{gini}
\alias{gini}
\title{GINI Coefficient}
\usage{
gini(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels. Can be \code{numeric, character, or factor}}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the GINI coefficient for a binary classifcation model
}
\examples{
data(testDF)
glmModel <- glm(y ~ ., data = testDF, family="binomial")
Preds <- predict(glmModel, type = 'response')

gini(testDF$y, Preds)

}

