% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{kappa}
\alias{kappa}
\title{kappa statistic}
\usage{
kappa(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{cutoff}{A cutoff for the predicted values}
}
\description{
Calculates kappa statistic. Currently build to handle binary values in \code{actual} vector.
}
