% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{logLoss}
\alias{logLoss}
\title{Log Loss}
\usage{
logLoss(actual, predicted, distribution = "binomial")
}
\arguments{
\item{actual}{a binary vector of the labels}

\item{predicted}{a vector of predicted values}

\item{distribution}{the distribution of the loss function needed \code{binomial, poisson}}
}

