% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape.R
\name{mape}
\alias{mape}
\alias{mape.default}
\alias{mape.gbm}
\alias{mape.glm}
\alias{mape.glmerMod}
\alias{mape.randomForest}
\alias{mape.rpart}
\title{Mean absolute percent error}
\usage{
mape(...)

\method{mape}{default}(actual, predicted, ...)

\method{mape}{glm}(modelObject, ...)

\method{mape}{randomForest}(modelObject, ...)

\method{mape}{glmerMod}(modelObject, ...)

\method{mape}{gbm}(modelObject, ...)

\method{mape}{rpart}(modelObject, ...)
}
\arguments{
\item{actual}{A vector actual responses}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{glm, randomForest, glmerMod, gbm}}

\item{\dots}{additional parameters to be passed the the s3 methods}
}
\description{
Calculates the mean absolute percent error
}
\examples{

 actual <- 1:3
 predicted <- actual / 2

 mape(actual,predicted)

}

