% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{mcc}
\alias{mcc}
\title{Matthews Correlation Coefficient}
\usage{
mcc(actual, predicted, cutoff)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{cutoff}{A cutoff for the predicted values}
}
\description{
Calculates the Matthews Correlation Coefficient
}
