% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{mlogLoss}
\alias{mlogLoss}
\title{Multiclass Log Loss}
\usage{
mlogLoss(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels. Can be \code{numeric, character, or factor}}

\item{predicted}{matrix of predicted values. Can be \code{matrix, data.frame}}
}
\description{
Calculated the multi-class log loss
}

