% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{msle}
\alias{msle}
\title{Mean Squared Log Error}
\usage{
msle(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the mean square log error
}

