% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{npv}
\alias{npv}
\title{Negative Predictive Value}
\usage{
npv(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{cutoff}{A cutoff for the predicted values

data(testDF)
glmModel <- glm(y ~ ., data = testDF, family="binomial")
Preds <- predict(glmModel, type = 'response')

npv(testDF$y, Preds, cutoff = 0)}
}
\description{
True Negatives / (True Negatives + False Negatives)
}

