% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{rmse}
\alias{rmse}
\title{Root-Mean Square Error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values

data(testDF)
glmModel <- glm(y ~ ., data = testDF, family="binomial")
Preds <- predict(glmModel, type = 'response')

rmse(testDF$y, Preds)}
}
\description{
Calculates the root mean square error
}

