% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{rmsle}
\alias{rmsle}
\title{Root Mean Squared Log Error}
\usage{
rmsle(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the mean square log error
}

