% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brier.R
\name{brier}
\alias{brier}
\alias{brier.default}
\alias{brier.gbm}
\alias{brier.glm}
\alias{brier.glmerMod}
\alias{brier.randomForest}
\alias{brier.rpart}
\title{Brier Score}
\usage{
brier(...)

\method{brier}{default}(actual, predicted, ...)

\method{brier}{glm}(modelObject, ...)

\method{brier}{randomForest}(modelObject, ...)

\method{brier}{glmerMod}(modelObject, ...)

\method{brier}{gbm}(modelObject, ...)

\method{brier}{rpart}(modelObject, ...)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{glm, randomForest, glmerMod, gbm}}

\item{\dots}{additional parameters to be passed the the s3 methods}
}
\description{
Calculates the Brier score
}

