% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ce.R
\name{ce}
\alias{ce}
\alias{ce.default}
\alias{ce.gbm}
\alias{ce.glm}
\alias{ce.glmerMod}
\alias{ce.lm}
\alias{ce.randomForest}
\alias{ce.rpart}
\title{Classification error}
\usage{
ce(...)

\method{ce}{default}(actual, predicted, ...)

\method{ce}{lm}(modelObject, ...)

\method{ce}{glm}(modelObject, ...)

\method{ce}{randomForest}(modelObject, ...)

\method{ce}{glmerMod}(modelObject, ...)

\method{ce}{gbm}(modelObject, ...)

\method{ce}{rpart}(modelObject, ...)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{lm, glm, randomForest, glmerMod, gbm, rpart}}

\item{\dots}{additional parameters to be passed the the s3 methods}
}
\description{
Calculates the classification error
}

