% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{ce}
\alias{ce}
\title{Classification error}
\usage{
ce(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the classification error
}

