% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\alias{mae.default}
\alias{mae.glm}
\alias{mae.randomForest}
\alias{mae.glmerMod}
\alias{mae.gbm}
\alias{mae.rpart}
\title{Mean absolute error}
\usage{
mae(...)

\method{mae}{default}(actual, predicted, ...)

\method{mae}{glm}(modelObject, ...)

\method{mae}{randomForest}(modelObject, ...)

\method{mae}{glmerMod}(modelObject, ...)

\method{mae}{gbm}(modelObject, ...)

\method{mae}{rpart}(modelObject, ...)
}
\arguments{
\item{\dots}{additional parameters to be passed the the s3 methods}

\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{glm, randomForest, glmerMod, gbm}}
}
\description{
Calculates the mean absolute error
}
