% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{mae}
\alias{mae}
\title{Mean absolute error}
\usage{
mae(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the mean absolute error
}

