% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\alias{mse.default}
\alias{mse.lm}
\alias{mse.glm}
\title{Mean Square Error}
\usage{
mse(...)

\method{mse}{default}(actual, predicted, ...)

\method{mse}{lm}(modelObject, ...)

\method{mse}{glm}(modelObject, ...)
}
\arguments{
\item{\dots}{additional parameters to be passed the the s3 methods}

\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{lm}}
}
\description{
Calculates the mean square error
}
\examples{
data(testDF)
glmModel <- glm(y ~ ., data = testDF, family="binomial")
Preds <- predict(glmModel, type = 'response')

mse(testDF$y, Preds)

}
