% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{mse}
\alias{mse}
\title{Mean Square Error}
\usage{
mse(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the mean square error
}
\examples{
data(testDF)
glmModel <- glm(y ~ ., data=df, family="binomial")
Preds <- predict(glmModel, type = 'response')

mse(testDF$y, Preds)

}

