% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msle.R
\name{msle}
\alias{msle}
\alias{msle.default}
\alias{msle.lm}
\alias{msle.glm}
\alias{msle.randomForest}
\alias{msle.glmerMod}
\alias{msle.gbm}
\alias{msle.rpart}
\title{Mean Squared Log Error}
\usage{
msle(...)

\method{msle}{default}(actual, predicted, ...)

\method{msle}{lm}(modelObject, ...)

\method{msle}{glm}(modelObject, ...)

\method{msle}{randomForest}(modelObject, ...)

\method{msle}{glmerMod}(modelObject, ...)

\method{msle}{gbm}(modelObject, ...)

\method{msle}{rpart}(modelObject, ...)
}
\arguments{
\item{\dots}{additional parameters to be passed the the s3 methods}

\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{glm, randomForest, glmerMod, gbm}}
}
\description{
Calculates the mean square log error
}
