% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(actual, predicted, cutoff = 0.5)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{cutoff}{A cutoff for the predicted values}
}
\description{
True Positives / (True Positives + False Negatives)
}

