% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMetrics.R
\name{rmse}
\alias{rmse}
\title{Root-Mean Square Error}
\usage{
rmse(actual, predicted)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the root mean square error
}

