% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{rmse}
\alias{rmse}
\alias{rmse.default}
\alias{rmse.lm}
\title{Root-Mean Square Error}
\usage{
rmse(...)

\method{rmse}{default}(actual, predicted)

\method{rmse}{lm}(object, ...)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}
}
\description{
Calculates the root mean square error
}
\examples{
data(testDF)
glmModel <- glm(y ~ ., data = testDF, family="binomial")
Preds <- predict(glmModel, type = 'response')

rmse(testDF$y, Preds)

}

