% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsle.R
\name{rmsle}
\alias{rmsle}
\alias{rmsle.default}
\alias{rmsle.gbm}
\alias{rmsle.glm}
\alias{rmsle.glmerMod}
\alias{rmsle.lm}
\alias{rmsle.randomForest}
\alias{rmsle.rpart}
\title{Root Mean Squared Log Error}
\usage{
rmsle(...)

\method{rmsle}{default}(actual, predicted)

\method{rmsle}{lm}(modelObject, ...)

\method{rmsle}{glm}(modelObject, ...)

\method{rmsle}{randomForest}(modelObject, ...)

\method{rmsle}{glmerMod}(modelObject, ...)

\method{rmsle}{gbm}(modelObject, ...)

\method{rmsle}{rpart}(modelObject, ...)
}
\arguments{
\item{actual}{A vector of the labels}

\item{predicted}{A vector of predicted values}

\item{modelObject}{the model object. Currently supported \code{glm, randomForest, glmerMod, gbm}}

\item{\dots}{additional parameters to be passed the the s3 methods}
}
\description{
Calculates the mean square log error
}

